finish off adding analyzers
