--- samples/xamarin-forms/Cinephile/Droid/AkavacheSqliteLinkerOverride.cs
@@ -1,25 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System;
-using Akavache.Sqlite3;
-
-// Note: This class file is *required* for iOS to work correctly, and is 
-// also a good idea for Android if you enable "Link All Assemblies".
-namespace Cinephile.Droid
-{
-    [Preserve]
-    public static class LinkerPreserve
-    {
-        static LinkerPreserve()
-        {
-            throw new Exception(typeof(SQLitePersistentBlobCache).FullName);
-        }
-    }
-
-
-    public class PreserveAttribute : Attribute
-    {
-    }
-}
