+++ samples/xamarin-forms/Cinephile/Droid/Cinephile.Droid.csproj
@@ -57,10 +57,11 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="LinkerPreserve.cs" />
     <Compile Include="MainActivity.cs" />
     <Compile Include="Resources\Resource.designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="AkavacheSqliteLinkerOverride.cs" />
+    <Compile Include="PreserveAttribute.cs" />
     <Compile Include="SplashActivity.cs" />
   </ItemGroup>
   <ItemGroup>
