+++ samples/xamarin-forms/Cinephile/Droid/SplashActivity.cs
@@ -1,20 +1,19 @@
-﻿
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿// Copyright (c) 2019 .NET Foundation and Contributors. All rights reserved.
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for full license information.
 
 using Android.App;
 using Android.Content;
 using Android.Content.PM;
 using Android.OS;
-using Android.Runtime;
 using Android.Support.V7.App;
-using Android.Views;
-using Android.Widget;
 
 namespace Cinephile.Droid
 {
+    /// <summary>
+    /// An activity for the splash screen.
+    /// </summary>
     [Activity(
         Label = "Cinephile",
         Icon = "@drawable/icon",
@@ -27,11 +26,7 @@ namespace Cinephile.Droid
     {
         private static readonly string TAG = "X:" + typeof(SplashActivity).Name;
 
-        public override void OnCreate(Bundle savedInstanceState, PersistableBundle persistentState)
-        {
-            base.OnCreate(savedInstanceState, persistentState);
-        }
-
+        /// <inheritdoc/>
         protected override void OnResume()
         {
             base.OnResume();
