+++ samples/xamarin-forms/Cinephile/iOS/AppDelegate.cs
@@ -1,20 +1,21 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+﻿// Copyright (c) 2019 .NET Foundation and Contributors. All rights reserved.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
+// See the LICENSE file in the project root for full license information.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using Akavache;
-using FFImageLoading.Forms.Touch;
 using Foundation;
 using UIKit;
 
 namespace Cinephile.iOS
 {
+    /// <summary>
+    /// The application delegate for the application.
+    /// </summary>
     [Register("AppDelegate")]
     public partial class AppDelegate : global::Xamarin.Forms.Platform.iOS.FormsApplicationDelegate
     {
+        /// <inheritdoc/>
         public override bool FinishedLaunching(UIApplication app, NSDictionary options)
         {
             UIApplication.SharedApplication.SetStatusBarStyle(UIStatusBarStyle.LightContent, false);
@@ -29,6 +30,7 @@ namespace Cinephile.iOS
             return base.FinishedLaunching(app, options);
         }
 
+        /// <inheritdoc/>
         public override void WillTerminate(UIApplication uiApplication)
         {
             BlobCache.Shutdown();
