+++ samples/xamarin-forms/Cinephile/iOS/Application.cs
@@ -1,0 +1,26 @@
+﻿// Copyright (c) 2019 .NET Foundation and Contributors. All rights reserved.
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for full license information.
+
+using UIKit;
+
+namespace Cinephile.iOS
+{
+    /// <summary>
+    /// Hosts the main application entry point for the application.
+    /// </summary>
+    public static class Application
+    {
+        /// <summary>
+        /// The main entry point for the application.
+        /// </summary>
+        /// <param name="args">Arguments passed to the application.</param>
+        private static void Main(string[] args)
+        {
+            // if you want to use a different Application Delegate class from "AppDelegate"
+            // you can specify it here.
+            UIApplication.Main(args, null, "AppDelegate");
+        }
+    }
+}
