+++ samples/xamarin-forms/Cinephile/iOS/Cinephile.iOS.csproj
@@ -30,6 +30,8 @@
     <MtouchDebug>true</MtouchDebug>
     <MtouchFastDev>true</MtouchFastDev>
     <OptimizePNGs>false</OptimizePNGs>
+    <NoWarn>;1591;1701;1702;1705;VSX1000;SA1300</NoWarn>
+    <DocumentationFile>bin\iPhoneSimulator\Debug\Cinephile.iOS.xml</DocumentationFile>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|iPhone' ">
     <DebugType>none</DebugType>
@@ -53,6 +55,8 @@
     <MtouchLink>None</MtouchLink>
     <MtouchArch>x86_64</MtouchArch>
     <ConsolePause>false</ConsolePause>
+    <NoWarn>;1591;1701;1702;1705;VSX1000;SA1300</NoWarn>
+    <DocumentationFile>bin\iPhoneSimulator\Release\Cinephile.iOS.xml</DocumentationFile>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|iPhone' ">
     <DebugSymbols>true</DebugSymbols>
@@ -151,9 +155,7 @@
       <Visible>false</Visible>
     </ImageAsset>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Resources\" />
-  </ItemGroup>
+  <ItemGroup />
   <ItemGroup>
     <InterfaceDefinition Include="LaunchScreen.storyboard" />
   </ItemGroup>
@@ -164,9 +166,11 @@
     <None Include="iOS.licenseheader" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Main.cs" />
+    <Compile Include="LinkerPreserve.cs" />
+    <Compile Include="Application.cs" />
     <Compile Include="AppDelegate.cs" />
-    <Compile Include="AkavacheSqliteLinkerOverride.cs" />
+    <Compile Include="PreserveAttribute.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Xamarin.Forms" Version="3.4.0.1029999" />
