+++ samples/xamarin-forms/Cinephile/iOS/Properties/AssemblyInfo.cs
@@ -7,7 +7,7 @@ using System.Reflection;
 
 // Information about this assembly is defined by the following attributes.
 // Change them to the values specific to your project.
-[assembly: AssemblyTitle("Cinephile.Droid")]
+[assembly: AssemblyTitle("Cinephile.iOS")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
